[BLINK-18562120] RM will encounter integer overflow when network buffer is transformed from MB to bytes
Summary: RM will encounter integer overflow when network buffer is transformed from MB to bytes

Reviewers: 五藏

Reviewed By: 五藏

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D814694

+++ flink-kubernetes/src/main/java/org/apache/flink/kubernetes/runtime/clusterframework/KubernetesResourceManager.java
@@ -529,7 +529,7 @@ public class KubernetesResourceManager extends ResourceManager<KubernetesWorkerN
 			additionalEnvs.put(Constants.ENV_TM_PROCESS_NETTY_MEMORY,
 				String.valueOf(taskManagerResource.getTaskManagerNettyMemorySizeMB()));
 
-			long networkBufBytes = taskManagerResource.getNetworkMemorySize() << 20;
+			long networkBufBytes = ((long) taskManagerResource.getNetworkMemorySize()) << 20;
 			additionalEnvs.put(Constants.ENV_TM_NETWORK_BUFFERS_MEMORY_MIN, String.valueOf(networkBufBytes));
 			additionalEnvs.put(Constants.ENV_TM_NETWORK_BUFFERS_MEMORY_MAX, String.valueOf(networkBufBytes));
 			log.debug("Task manager environments: " + additionalEnvs);
